refactor(models): consolidate classroom models into single module
Consolidate Assignment, Question, ClassroomLink, Submission, Response,
and Feedback models from separate files into campus/model/classroom.py,
following the pattern established by timetable.py.

Changes:
- Create campus/model/classroom.py with all 6 classroom-related models
- Update campus/model/__init__.py to import from classroom module
- Delete campus/model/assignment.py and campus/model/submission.py

This improves domain coherence by keeping all classroom models together,
making relationships and dependencies more visible.
